requirements.txt
pandas
numpy
scikit-learn
transformers
datasets
torchmetrics
setuptools
flash-attention 
accelerate
panel
holoviews
plotly

# pip freeze 

# aiohappyeyeballs==2.4.4
# aiohttp==3.11.10
# aiosignal==1.3.1
# annotated-types==0.7.0
# attrs==24.2.0
# autoflake==2.3.1
# black==24.10.0
# blinker==1.9.0
# Brotli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_27zk0eqdh0/croot/brotli-split_1714483157007/work
# cachelib==0.9.0
# certifi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_0ecs_ou28u/croot/certifi_1725551740278/work/certifi
# charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
# click==8.1.7
# dash==2.18.2
# dash-bootstrap-components==1.6.0
# dash-core-components==2.0.0
# dash-html-components==2.0.0
# dash-mantine-components==0.12.1
# dash-table==5.0.0
# dash_ag_grid==31.3.0
# datasets==3.1.0
# dill==0.3.8
# filelock @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_d3quwmvouf/croot/filelock_1700591194006/work
# Flask==3.0.3
# Flask-Caching==2.3.0
# frozenlist==1.5.0
# fsspec==2024.9.0
# gmpy2 @ file:///Users/cbousseau/work/recipes/ci_py311/gmpy2_1677937751357/work
# huggingface-hub==0.26.5
# idna @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a12xpo84t2/croot/idna_1714398852854/work
# importlib_metadata==8.5.0
# itsdangerous==2.2.0
# Jinja2 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b15nuwux5r/croot/jinja2_1730902833938/work
# joblib==1.4.2
# lightning-utilities==0.11.9
# MarkupSafe @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a84ni4pci8/croot/markupsafe_1704206002077/work
# mpmath @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_17iu6a8a3m/croot/mpmath_1690848269369/work
# multidict==6.1.0
# multiprocess==0.70.16
# mypy-extensions==1.0.0
# nest-asyncio==1.6.0
# networkx @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_797xfkx_y7/croot/networkx_1717597499877/work
# numpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_f2amvjkczw/croot/numpy_and_numpy_base_1725470317860/work/dist/numpy-2.0.1-cp311-cp311-macosx_11_0_arm64.whl#sha256=1faed35ced6c9bbc9c345982a6f5c2ad89fc0b526315b56fc2cdb4b4ec6f9c42
# packaging==24.2
# pandas==2.2.3
# pathspec==0.12.1
# pillow @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_92egn12how/croot/pillow_1731594702114/work
# platformdirs==4.3.6
# plotly==5.24.1
# propcache==0.2.1
# pyarrow==18.1.0
# pydantic==2.10.3
# pydantic_core==2.27.1
# pyflakes==3.2.0
# PySocks @ file:///Users/cbousseau/work/recipes/ci_py311/pysocks_1677906386870/work
# python-dateutil==2.9.0.post0
# pytz==2024.2
# PyYAML @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_faoex52hrr/croot/pyyaml_1728657970485/work
# regex==2024.11.6
# requests @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ee45nsd33z/croot/requests_1730999134038/work
# retrying==1.3.4
# safetensors==0.4.5
# scikit-learn==1.5.2
# scipy==1.14.1
# six==1.17.0
# sympy==1.13.1
# tenacity==9.0.0
# threadpoolctl==3.5.0
# tokenizers==0.21.0
# torch==2.6.0.dev20241112
# torchaudio==2.5.0.dev20241118
# torchmetrics==1.6.0
# torchvision==0.20.0.dev20241118
# tqdm==4.67.1
# transformers==4.47.0
# typing_extensions==4.12.2
# tzdata==2024.2
# urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_06_m8gdsy6/croot/urllib3_1727769822458/work
# vizro==0.1.29
# Werkzeug==3.0.6
# wrapt==1.17.0
# xxhash==3.5.0
# yarl==1.18.3
# zipp==3.21.0
